Add support for allocating the LZMA2 dictionary dynamically.
Previously the dictionary was preallocated at initialization
time, which is useful since the decoder cannot then later run
out of memory, but in several cases it is just an annoying
limitation.

It is now possible to enable only the needed operation mode(s)
at build time, which saves a few bytes in code size if only
one or two modes are actually needed. Bigger savings would be
possible especially in single-call mode, but I'll think about
that later.

This commit changes the API by adding the mode argument to
xz_dec_init(). A new return value (XZ_MEM_ERROR) was also added,
but it is used only in the new XZ_DYNALLOC operation mode.