Update linux/lib/xz/Kconfig from Linux 3.9.
This drops "if EXPERT" from BCJ filter configuration, and the
default configuration enables only the BCJ filter(s) that are
likely needed on the target arch: x86 BCJ filter is enabled
on x86, PowerPC filter on PowerPC, and so on.

Patches do this were made by Florian Fainelli, and a typo was
fixed by Paul Bolle.