Minor README updates.
This rewords a few minor things. Additionally, the table macros was
updated to include the forgotten XZ_DEC_ARM64 and to note that the ARM
based architectures can use the filters in little or big endian mode, as
long as the instructions are in little endian mode. Modern ARM
architectures use little endian instruction encoding, even if other data
is encoded big endian.