userspace/boottest.c: Make it build with the new __decompress API.
The __decompress() function takes a new extra argument to specify
output buffer size. Some decompressors in Linux use it, some don't.
Those that don't, assume that output buffer is SIZE_MAX. It's not
a security problem because the code is used for decompressing the
kernel image only but it was still one of odd things in the API design.

Buffer sizes are now long instead of int (but not size_t)
so that initramfs larger than 2 GiB works too.