userspace/Makefile: Switch from -std=gnu89 to -std=gnu11.
Linux moved from -std=gnu89 to -std=gnu11 in 2022 so now we need
much fewer extensions than in the past (long long, stdint.h, and
stdbool.h are standard C99 and later).

The option is -Wdeclaration-after-statement there for now because
mixed declarations and statements aren't used at the moment
(even though I like those) but it's not worth it to change those
at this point.

These are the known things that don't conform to C99/C11:

  - Null pointer + 0 is assumed to be well defined:

        char *p = NULL; p += 0; assert(p == NULL);

    It's required if a pointer in struct xz_buf is NULL
    and thus the matching foo_size is 0.

  - LZMA decoder accesses arrays inside structures in a way that
    goes beyond the array bounds (while still staying within
    the structure). The initialization thing could be neatly
    cleaned up with C11's anonymous union.