Add linux/MAINTAINERS.
It includes only the lines that are relevant to this project.
Having the file here works as a reminder that those lines need
to be updated in Linux if the information becomes outdated.

Note that the file list (F: entries) must be in case-insensitive
alphabetical order to make checkpatch.pl happy.