Override default bug report template.
This provides an extra note to mention if the version is from the Linux
kernel or regular XZ Embedded. There are a few minor differences that
could help us determine what the problem is.

Also, the reporter may not be able to determine the version, so that
field is not required.