Avoid "eval" when setting XZ_VERSION.
The xz man page has the example that uses "eval". Due to the
XZ Utils backdoor incident, the use of "eval" it understandably
makes people worried.

The new method is more robust in case XZ_OPT or XZ_DEFAULTS
contain options that usually shouldn't be there. For example,
with XZ_OPT=--help the "eval" method would execute the --help
output.