Fix kernel-doc formatting in xz.h
The opaque structs xz_dec and xz_dec_microlzma are declared in xz.h
but their definitions are in xz_dec_lzma2.c. The kernel-doc comments
can only be used for definitions so the comments in xz.h must be
regular comments to avoid errors from the scripts/kernel-doc tool
in Linux.

Add a few missing colons too.